Bulk Load CDK: CSV to its own toolkit; test stuff to test fixtures
