+++ airbyte-cdk/bulk/core/load/src/main/kotlin/io/airbyte/cdk/load/data/AirbyteValueIdentityMapper.kt
@@ -73,7 +73,6 @@ open class AirbyteValueIdentityMapper(
         path: List<String>
     ): AirbyteValue = value
 
-
     open fun mapArray(value: ArrayValue, schema: ArrayType, path: List<String>): AirbyteValue {
         return ArrayValue(
             value.values.mapIndexed { index, element ->
@@ -81,7 +80,6 @@ open class AirbyteValueIdentityMapper(
             }
         )
     }
-
 
     open fun mapArrayWithoutSchema(
         value: ArrayValue,
@@ -113,4 +111,6 @@ open class AirbyteValueIdentityMapper(
         value
 
     open fun mapNull(path: List<String>): AirbyteValue = NullValue
+
+    open fun mapUnknown(value: UnknownValue, path: List<String>): AirbyteValue = value
 }
