+++ airbyte-cdk/bulk/toolkits/load-csv/src/main/kotlin/io/airbyte/cdk/load/file/csv/CSVWriter.kt
@@ -2,21 +2,13 @@
  * Copyright (c) 2024 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.cdk.load.data
+package io.airbyte.cdk.load.file.csv
 
+import io.airbyte.cdk.load.data.ObjectType
+import io.airbyte.cdk.load.data.csv.toCsvHeader
 import java.io.Writer
 import org.apache.commons.csv.CSVFormat
 import org.apache.commons.csv.CSVPrinter
-
-class AirbyteTypeToCsvHeader {
-    fun convert(schema: ObjectType): Array<String> {
-        return schema.properties.map { it.key }.toTypedArray()
-    }
-}
-
-fun ObjectType.toCsvHeader(): Array<String> {
-    return AirbyteTypeToCsvHeader().convert(this)
-}
 
 fun ObjectType.toCsvPrinterWithHeader(writer: Writer): CSVPrinter =
     CSVFormat.Builder.create().setHeader(*toCsvHeader()).setAutoFlush(true).build().print(writer)
