+++ airbyte-cdk/bulk/toolkits/load-object-storage/build.gradle
@@ -6,9 +6,11 @@ dependencies {
     // and migrate them to their respective toolkits, so that these
     // dependencies can be removed.
     api project(':airbyte-cdk:bulk:toolkits:bulk-cdk-toolkit-load-avro')
+    api project(':airbyte-cdk:bulk:toolkits:bulk-cdk-toolkit-load-csv')
     api project(':airbyte-cdk:bulk:toolkits:bulk-cdk-toolkit-load-parquet')
 
-    api("org.apache.commons:commons-csv:1.10.0")
-
     testFixturesImplementation testFixtures(project(":airbyte-cdk:bulk:core:bulk-cdk-core-load"))
+    testFixturesImplementation testFixtures(project(":airbyte-cdk:bulk:toolkits:bulk-cdk-toolkit-load-avro"))
+    testFixturesImplementation testFixtures(project(":airbyte-cdk:bulk:toolkits:bulk-cdk-toolkit-load-csv"))
+    testFixturesImplementation testFixtures(project(":airbyte-cdk:bulk:toolkits:bulk-cdk-toolkit-load-parquet"))
 }
